+++ model_navigator/api/config.py
@@ -503,6 +503,9 @@ DEFAULT_TENSORRT_PRECISION = (
 )
 DEFAULT_TENSORRT_PRECISION_MODE = TensorRTPrecisionMode.HIERARCHY
 
+TensorRTPrecisionType = Union[Union[str, TensorRTPrecision], Tuple[Union[str, TensorRTPrecision], ...]]
+TensorRTPrecisionModeType = Union[str, TensorRTPrecisionMode]
+
 
 class CustomConfig(abc.ABC):
     """Base class used for custom configs. Input for Model Navigator `optimize` method."""
@@ -548,10 +551,8 @@ class CustomConfigForTensorRT(CustomConfigForFormat):
 
     trt_profiles: Optional[List[TensorRTProfile]] = None
     trt_profile: Optional[TensorRTProfile] = None  # TODO: Remove before 1.0.0 release
-    precision: Union[Union[str, TensorRTPrecision], Tuple[Union[str, TensorRTPrecision], ...]] = (
-        DEFAULT_TENSORRT_PRECISION
-    )
-    precision_mode: Optional[Union[str, TensorRTPrecisionMode]] = DEFAULT_TENSORRT_PRECISION_MODE
+    precision: TensorRTPrecisionType = DEFAULT_TENSORRT_PRECISION
+    precision_mode: TensorRTPrecisionModeType = DEFAULT_TENSORRT_PRECISION_MODE
     max_workspace_size: Optional[int] = DEFAULT_MAX_WORKSPACE_SIZE
     run_max_batch_size_search: Optional[bool] = None  # TODO this parameter is currently not used